明細のKingaku/Taxをlong化し伝票合計の桁溢れを解消
発注入力の詳細画面を開くと「詳細取得失敗: Arithmetic operation resulted in
an overflow.」となっていた。UpdateTotals()の Sum(m => m.Su * m.Gedai) が
int*intのままでSum(int)のchecked加算に入るため、合計がintを超えて例外に
なっていた。

真因はデータ側にある。c06f141cで旧DB変換が上代金額/下代金額から
上代単価/下代単価へ修正されたが、既存DBは修正前のロジックで変換されたまま
再変換されていない。旧HC$tran_tori1には単価列と金額列が別に存在し、
cv10のJodai/Gedaiには金額が入っている(Tran13Hachu全18,966明細で
Gedai=Kingakuが100%成立)。このためSu*Gedaiが数量の二重計上になり、
発注Id=5000は明細1行で3,000*16,650,000=499億(int上限の23倍)に達していた。
データ再変換は別途対応する。

式自体は正しいため変更せず、算術の桁のみ修正した。
- Tran99Meisai.Kingaku / Tax を int -> long へ変更
  (明細はJmeisaiにJSONで入るためDBスキーマ変更は不要)
- TaxCalculatorのTaxAmountセッタからunchecked narrowing (int)value を除去
- Su*Jodai / Su*Gedai / Su*Tanka の乗算へ (long) キャストを付与
  long格納でも未キャストではint段階で溢れ、例外にならず静かに値が化ける
- 重複していたUpdateTotals相当4箇所とHhtProcessUpdateMapを同じ式に統一
- PointOfSaleServiceのJodaiTotal/GedaiTotalに欠落していた *Su を追加
  CreateLineがJodai=tanka(単価)を入れており単価の素合計は他6伝票と不一致
- Jodai/GedaiのOldTableCommentAttrを上代金額->上代単価へ訂正(誤記が誤解の温床)

+++ CvBase/TaxCalculator.cs
@@ -22,7 +22,7 @@ public interface ITaxMeisaiLine {
 
 /// <summary>
 /// <see cref="Tran99Meisai"/> の <see cref="ITaxMeisaiLine"/> 実装。
-/// <c>Kingaku</c>/<c>Tax</c> が int のため、long⇔int の変換をここで吸収する。
+/// <c>Kingaku</c>/<c>Tax</c> ともに long のため、そのまま委譲する。
 /// </summary>
 public sealed partial class Tran99Meisai : ITaxMeisaiLine {
 	long ITaxMeisaiLine.TaxKingaku => Kingaku;
@@ -36,7 +36,7 @@ public sealed partial class Tran99Meisai : ITaxMeisaiLine {
 	}
 	long ITaxMeisaiLine.TaxAmount {
 		get => Tax;
-		set => Tax = (int)value;
+		set => Tax = value;
 	}
 }
