+++ CvWpfclient/Helpers/ViewModels/BaseIdoInputViewModel.cs
@@ -302,7 +302,7 @@ order by h.DenDay desc, h.Id desc, cast({M}'$.No') as int)
 			}
 
 			row.No = nextNo++;
-			row.Kingaku = row.Su * row.Tanka;
+			row.Kingaku = (long)row.Su * row.Tanka;
 			row.PropertyChanged += OnMeisaiPropertyChanged;
 			EditMeisai.Add(row);
 			SelectedMeisai = row;
@@ -371,7 +371,7 @@ order by h.DenDay desc, h.Id desc, cast({M}'$.No') as int)
 		meisai.Code_Siz = row.Source.Code_Siz;
 		meisai.Mei_Siz = row.Source.Mei_Siz;
 		meisai.Su = row.Su;
-		meisai.Kingaku = meisai.Su * meisai.Tanka;
+		meisai.Kingaku = (long)meisai.Su * meisai.Tanka;
 		meisai.JanCode = row.Source.Jan1;
 	}
